ProductOption 컨트롤러 통합
